+++ instagram_private_api/__init__.py
@@ -4,10 +4,12 @@ from .client import Client
 from .compatpatch import ClientCompatPatch
 from .errors import (
     ClientError, ClientLoginError, ClientLoginRequiredError,
-    ClientCookieExpiredError, ClientThrottledError, ClientConnectionError
+    ClientCookieExpiredError, ClientThrottledError, ClientConnectionError,
+    ClientCheckpointRequiredError, ClientChallengeRequiredError,
+    ClientSentryBlockError,
 )
 from .endpoints.upload import MediaRatios
 from .endpoints.common import MediaTypes
 
 
-__version__ = '1.3.6'
+__version__ = '1.4.1'