+++ instagram_private_api/endpoints/media.py
@@ -71,7 +71,12 @@ class MediaEndpointsMixin(object):
         :return:
         """
         endpoint = 'media/{media_id!s}/comments/'.format(**{'media_id': media_id})
-        res = self._call_api(endpoint, query=kwargs)
+        query = {
+            'can_support_threading': 'true'
+        }
+        if kwargs:
+            query.update(kwargs)
+        res = self._call_api(endpoint, query=query)
 
         if self.auto_patch:
             [ClientCompatPatch.comment(c, drop_incompat_keys=self.drop_incompat_keys)
@@ -107,6 +112,51 @@ class MediaEndpointsMixin(object):
              for c in comments]
 
         return sorted(comments, key=lambda k: k['created_at_utc'], reverse=reverse)
+
+    def comment_replies(self, media_id, comment_id, **kwargs):
+        """
+        Get comment replies. Fixed at 20 replies returned per page.
+        Check for 'has_more_tail_child_comments', 'next_max_child_cursor' to determine
+        if there are more replies to page through.
+
+        :param media_id: Media id
+        :param comment_id: Comment id
+        :param kwargs:
+            **max_id**: For pagination
+        :return:
+        """
+        endpoint = 'media/{media_id!s}/comments/{comment_id!s}/child_comments/'.format(
+            **{'media_id': media_id, 'comment_id': comment_id})
+        res = self._call_api(endpoint, query=kwargs)
+
+        if self.auto_patch:
+            [ClientCompatPatch.comment(c, drop_incompat_keys=self.drop_incompat_keys)
+             for c in res.get('child_comments', [])]
+            ClientCompatPatch.comment(res.get('parent_comment'))
+        return res
+
+    def comment_inline_replies(self, media_id, comment_id, max_id, **kwargs):
+        """
+        Get inline comment replies.
+        Check for 'next_max_child_cursor' from ``media_comments()``
+        to determine if there are inline comment replies to retrieve.
+
+        :param media_id: Media id
+        :param comment_id: Comment id
+        :param max_id: The comment's 'next_max_child_cursor' value from``media_comments()``
+        :return:
+        """
+        endpoint = 'media/{media_id!s}/comments/{comment_id!s}/inline_child_comments/'.format(
+            **{'media_id': media_id, 'comment_id': comment_id})
+        query = {'max_id': max_id}
+        if kwargs:
+            query.update(kwargs)
+        res = self._call_api(endpoint, query=query)
+        if self.auto_patch:
+            [ClientCompatPatch.comment(c, drop_incompat_keys=self.drop_incompat_keys)
+             for c in res.get('child_comments', [])]
+            ClientCompatPatch.comment(res.get('parent_comment'))
+        return res
 
     def edit_media(self, media_id, caption, usertags=None):
         """
@@ -514,3 +564,16 @@ class MediaEndpointsMixin(object):
         :param media_type: One of :attr:`MediaTypes.PHOTO`, :attr:`MediaTypes.VIDEO`, or :attr:`MediaTypes.CAROUSEL`
         """
         return self.media_only_me(media_id, media_type, undo=True)
+
+    def story_viewers(self, story_pk, **kwargs):
+        """
+        Get list of story viewers
+
+        :param story_pk: Story media's PK identifier, e.g. "1700000123"
+        :param kwargs:
+            **max_id**: For pagination
+        :return:
+        """
+        endpoint = 'media/{story_pk!s}/list_reel_media_viewer/'.format(
+            story_pk=story_pk)
+        return self._call_api(endpoint, query=kwargs)