+++ instagram_web_api/__init__.py
@@ -2,8 +2,11 @@
 
 from .client import Client
 from .compatpatch import ClientCompatPatch
-from .errors import ClientError, ClientLoginError, ClientCookieExpiredError
+from .errors import (
+    ClientError, ClientLoginError, ClientCookieExpiredError,
+    ClientConnectionError
+)
 from .common import ClientDeprecationWarning
 
 
-__version__ = '1.3.6'
+__version__ = '1.4.1'