+++ tests/common.py
@@ -12,25 +12,31 @@ try:
     from instagram_private_api import (
         __version__, Client, ClientError, ClientLoginError,
         ClientCookieExpiredError, ClientThrottledError, ClientCompatPatch,
-        ClientLoginRequiredError, MediaTypes)
+        ClientLoginRequiredError, MediaTypes,
+        ClientSentryBlockError, ClientCheckpointRequiredError,
+        ClientChallengeRequiredError)
     from instagram_private_api.utils import (
         InstagramID, gen_user_breadcrumb,
         max_chunk_size_generator, max_chunk_count_generator, get_file_size,
         ig_chunk_generator
     )   # noqa
     from instagram_private_api.constants import Constants
+    from instagram_private_api.compat import compat_urllib_parse
 except ImportError:
     sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
     from instagram_private_api import (
         __version__, Client, ClientError, ClientLoginError,
         ClientCookieExpiredError, ClientThrottledError, ClientCompatPatch,
-        ClientLoginRequiredError, MediaTypes)
+        ClientLoginRequiredError, MediaTypes,
+        ClientSentryBlockError, ClientCheckpointRequiredError,
+        ClientChallengeRequiredError)
     from instagram_private_api.utils import (
         InstagramID, gen_user_breadcrumb,
         max_chunk_size_generator, max_chunk_count_generator, get_file_size,
         ig_chunk_generator
     )   # noqa
     from instagram_private_api.constants import Constants
+    from instagram_private_api.compat import compat_urllib_parse
 
 try:
     from instagram_web_api import (
@@ -126,9 +132,13 @@ class WebApiTestBase(unittest.TestCase):
 class MockResponse(object):
     """A mock class to emulate api responses."""
 
-    def __init__(self, code=200, content_type=''):
+    def __init__(self, code=200, content_type='', body=''):
         self.code = 200
         self.content_type = content_type
+        self.body = body
 
     def info(self):
         return {'Content-Type': self.content_type}
+
+    def read(self):
+        return self.body.encode('utf8')