+++ tests/private/accounts.py
@@ -87,7 +87,19 @@ class AccountTests(ApiTestBase):
             {
                 'name': 'test_change_profile_picture_mock',
                 'test': AccountTests('test_change_profile_picture_mock', api)
-            }
+            },
+            {
+                'name': 'test_presence_status',
+                'test': AccountTests('test_presence_status', api)
+            },
+            {
+                'name': 'test_enable_presence_status_mock',
+                'test': AccountTests('test_enable_presence_status_mock', api)
+            },
+            {
+                'name': 'test_disable_presence_status_mock',
+                'test': AccountTests('test_disable_presence_status_mock', api)
+            },
         ]
 
     @unittest.skip('Unwise to run frequently.')
@@ -152,13 +164,13 @@ class AccountTests(ApiTestBase):
                 '',     # Test 1
                 '',     # Test 1
                 '',     # Test 2
-                compat_urllib_error.HTTPError(      # Test 2
-                    self.api.api_url, 500, 'Internal Server Error', {},
-                    BytesIO('Internal Server Error'.encode('utf-8'))),
+                ClientError(        # Test 2
+                    'Internal Server Error', code=500,
+                    error_response='Internal Server Error'),
                 '',     # Test 3
-                compat_urllib_error.HTTPError(      # Test 2
-                    self.api.api_url, 400, 'Invalid', {},
-                    BytesIO('Invalid'.encode('utf-8'))),
+                ClientLoginError(       # Test 2
+                    'Invalid', code=400,
+                    error_response='Invalid'),
             ]
             read_response.return_value = json.dumps({'status': 'fail'})
 
@@ -173,9 +185,9 @@ class AccountTests(ApiTestBase):
             self.assertEqual(ce.exception.msg, 'Internal Server Error')
 
             # Test 3
-            with self.assertRaises(ClientLoginError) as cle:
+            with self.assertRaises(ClientError) as cle:
                 self.api.login()
-            self.assertEqual(cle.exception.msg, 'Unable to login: HTTP Error 400: Invalid')
+            self.assertEqual(cle.exception.msg, 'Invalid')
 
     def test_current_user(self):
         results = self.api.current_user()
@@ -378,3 +390,33 @@ class AccountTests(ApiTestBase):
             with self.assertRaises(ClientError) as he:
                 self.api.change_profile_picture(photo_data)
             self.assertEqual(str(he.exception), 'Internal Server Error')
+
+    def test_presence_status(self):
+        results = self.api.presence_status()
+        self.assertIn('disabled', results)
+
+    @compat_mock.patch('instagram_private_api.Client._call_api')
+    def test_enable_presence_status_mock(self, call_api):
+        call_api.return_value = {'status': 'ok'}
+        self.api.enable_presence_status()
+        call_api.assert_called_with(
+            'accounts/set_presence_disabled/',
+            params={
+                '_csrftoken': self.api.csrftoken,
+                '_uuid': self.api.uuid,
+                '_uid': self.api.authenticated_user_id,
+                'disabled': '0',
+            })
+
+    @compat_mock.patch('instagram_private_api.Client._call_api')
+    def test_disable_presence_status_mock(self, call_api):
+        call_api.return_value = {'status': 'ok'}
+        self.api.disable_presence_status()
+        call_api.assert_called_with(
+            'accounts/set_presence_disabled/',
+            params={
+                '_csrftoken': self.api.csrftoken,
+                '_uuid': self.api.uuid,
+                '_uid': self.api.authenticated_user_id,
+                'disabled': '1',
+            })