+++ tests/web/feed.py
@@ -16,16 +16,50 @@ class FeedTests(WebApiTestBase):
                 'name': 'test_location_feed',
                 'test': FeedTests('test_location_feed', api),
             },
+            {
+                'name': 'test_timeline_feed',
+                'test': FeedTests('test_timeline_feed', api),
+            },
+            {
+                'name': 'test_reels_tray',
+                'test': FeedTests('test_reels_tray', api),
+            },
+            {
+                'name': 'test_reels_feed',
+                'test': FeedTests('test_reels_feed', api),
+            },
         ]
 
     def test_tag_feed(self):
-        results = self.api.tag_feed('catsofinstagram')
-        self.assertIsNotNone(results.get('tag'))
-        self.assertGreater(len(results.get('tag', {}).get('media', {}).get('nodes', [])), 0)
-        self.assertGreater(len(results.get('tag', {}).get('top_posts', {}).get('nodes', [])), 0)
+        results = self.api.tag_feed('catsofinstagram').get('data', {})
+        self.assertIsNotNone(results.get('hashtag', {}).get('name'))
+        self.assertGreater(
+            len(results.get('hashtag', {}).get('edge_hashtag_to_media', {}).get('edges', [])), 0)
+        self.assertGreater(
+            len(results.get('hashtag', {}).get('edge_hashtag_to_top_posts', {}).get('edges', [])), 0)
 
     def test_location_feed(self):
-        results = self.api.location_feed('212988663')
-        self.assertIsNotNone(results.get('location'))
-        self.assertGreater(len(results.get('location', {}).get('media', {}).get('nodes', [])), 0)
-        self.assertGreater(len(results.get('location', {}).get('top_posts', {}).get('nodes', [])), 0)
+        results = self.api.location_feed('212988663').get('data', {})
+        self.assertIsNotNone(results.get('location', {}).get('name'))
+        self.assertGreater(
+            len(results.get('location', {}).get('edge_location_to_media', {}).get('edges', [])), 0)
+        self.assertGreater(
+            len(results.get('location', {}).get('edge_location_to_top_posts', {}).get('edges', [])), 0)
+
+    def test_timeline_feed(self):
+        results = self.api.timeline_feed().get('data', {})
+        self.assertIsNotNone(results.get('user', {}).get('username'))
+        self.assertGreater(
+            len(results.get('user', {}).get('edge_web_feed_timeline', {}).get('edges', [])), 0)
+
+    def test_reels_tray(self):
+        results = self.api.reels_tray().get('data', {})
+        self.assertGreater(
+            len(results.get('user', {}).get(
+                'feed_reels_tray', {}).get(
+                    'edge_reels_tray_to_reel', {}).get('edges', [])), 0)
+
+    def test_reels_feed(self):
+        results = self.api.reels_feed(['25025320']).get('data', {})
+        self.assertGreater(
+            len(results.get('reels_media')), 0)